+++ tests/test_doctests.py
@@ -23,8 +23,8 @@ def test_quantiphy():
     Quantity.reset_prefs()
     rv = doctest.testfile('../quantiphy.py', optionflags=doctest.ELLIPSIS)
     assert rv.failed == 0
-    assert rv.attempted == 100
-        # this target should be undated when the number of doctests change
+    assert rv.attempted == 119
+        # this target should be updated when the number of doctests change
 
 def test_manual():
     if sys.version_info < (3, 6):
@@ -33,7 +33,7 @@ def test_manual():
     Quantity.reset_prefs()
     expected_test_count = {
         '../doc/index.rst': 31,
-        '../doc/user.rst': 368,
+        '../doc/user.rst': 372,
         '../doc/api.rst': 0,
         '../doc/examples.rst': 36,
         '../doc/accessories.rst': 12,